backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
gunicorn==20.1.0
whitenoise==6.5.0
python-decouple==3.8
django-environ==0.10.0
cryptography==42.0.2
pillow==10.3.0
sqlparse==0.4.4
django-cors-headers==4.0.0
drf-yasg==1.21.5
